Be feat/question

+++ server/server-application/src/main/java/com/mzdevelopers/serverapplication/question/service/QuestionServiceImpl.java
@@ -1,14 +1,20 @@
 package com.mzdevelopers.serverapplication.question.service;
 
-import com.mzdevelopers.serverapplication.question.dto.StubAnswer;
+import com.mzdevelopers.serverapplication.question.stub.MemberStub;
+import com.mzdevelopers.serverapplication.question.stub.StubAnswer;
 import com.mzdevelopers.serverapplication.question.entity.Question;
+import com.mzdevelopers.serverapplication.question.entity.QuestionVote;
 import com.mzdevelopers.serverapplication.question.repository.QuestionRepository;
+import com.mzdevelopers.serverapplication.question.repository.QuestionVoteRepository;
+import com.mzdevelopers.serverapplication.question.stub.MemberStubRepository;
+import com.mzdevelopers.serverapplication.tag.dto.TagDto;
 import com.mzdevelopers.serverapplication.tag.entity.QuestionTag;
 import com.mzdevelopers.serverapplication.tag.entity.Tag;
 import com.mzdevelopers.serverapplication.tag.repository.QuestionTagRepository;
 import com.mzdevelopers.serverapplication.tag.repository.TagRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -26,9 +32,11 @@ public class QuestionServiceImpl implements QuestionService{
     private final QuestionRepository questionRepository;
     private final QuestionTagRepository questionTagRepository;
     private final TagRepository tagRepository;
+    private final QuestionVoteRepository questionVoteRepository;
+    private final MemberStubRepository memberStubRepository;
 
     @Override
-    public long creatQuestion(Question question, List<Long> tags) {
+    public long createQuestion(Question question, List<Long> tags) {
         Question savedQuestion = questionRepository.save(question);
 
         if (!tags.isEmpty()) {
@@ -45,9 +53,13 @@ public class QuestionServiceImpl implements QuestionService{
     }
 
     @Override
-    public Question getQuestion(long questionId) {
+    public Question getQuestion(long questionId, long memberId) {
         Question findQuestion = findByQuestionId(questionId);
-        System.out.println("getQuestion : " + findQuestion.getQuestionTags().get(0).toString());
+        save(); // stub -> delete
+        if(isRegisteredMember(memberId)){
+            findQuestion.increaseView();
+            questionRepository.save(findQuestion);
+        }
         return findQuestion;
     }
 
@@ -63,11 +75,59 @@ public class QuestionServiceImpl implements QuestionService{
     }
 
     @Override
-    public void deleteQuestion(long questionId) {
+    public void deleteQuestion(long questionId, long memberId) {
         Question findQuestion = findByQuestionId(questionId);
-        questionRepository.delete(findQuestion);
+        if(findQuestion.getMemberId() == memberId)
+            questionRepository.delete(findQuestion);
+        else
+            throw new IllegalArgumentException("삭제할 권한이 없습니다: " + memberId);
     }
 
+
+    // -------------------------------------------------------------- 질문 CRUD
+
+    @Override
+    public int votesCount(long questionId, long memberId) {
+        Question findQuestion = findByQuestionId(questionId);
+        MemberStub memberStub = save();
+        Optional<QuestionVote> optionalQuestionVote = questionVoteRepository.findByQuestionAndMemberStub(findQuestion, memberStub);
+        if (optionalQuestionVote.isEmpty()) {
+            QuestionVote questionVote = QuestionVote.builder().question(findQuestion).memberStub(memberStub).build();
+            findQuestion.updateVoteCount(true);
+            questionVoteRepository.save(questionVote);
+        } else {
+            QuestionVote findQuestionVote = optionalQuestionVote.get();
+            findQuestionVote.updateVote();
+            questionVoteRepository.save(findQuestionVote);
+            findQuestion.updateVoteCount(findQuestionVote.isQuestionVoted());
+        }
+        Question updatedQuestion = questionRepository.save(findQuestion);
+        return updatedQuestion.getVotesCount();
+    }
+    // ------------------------------------------------------------- 종아요 증가 or 감소
+
+    @Override
+    public List<Question> questionsListByAPI(int page, int size, String api) {
+        Pageable pageable = createPageable(page, size, "createdAt");
+        if (api.equals("votes")) {
+            pageable = createPageable(page, size, "votesCount");
+            return questionRepository.findAllByOrderByVotesCountDesc(pageable).getContent();
+        } else if(api.equals("solutions")) {
+            return questionRepository.findBySolutionStatusTrueOrderByCreatedAtDesc(pageable).getContent();
+        }else {
+            return questionRepository.findAll(pageable).getContent();
+        }
+    }
+
+    public Pageable createPageable(int page, int size, String property) {
+        return PageRequest.of(page, size, Sort.by(property).descending());
+    }
+
+    //--------------------------------------------------------------- vote, view, answer
+
+    public boolean isRegisteredMember(long memberId) {
+        return memberStubRepository.findById(memberId).isPresent();
+    }
     public Question findByQuestionId(Long questionId) {
         Optional<Question> findQuestion = questionRepository.findById(questionId);
         return findQuestion.orElseThrow(() -> new IllegalArgumentException("No Search Question: " + questionId));
@@ -82,21 +142,41 @@ public class QuestionServiceImpl implements QuestionService{
         return tagList;
     }
 
+    public List<TagDto> findByQuestionTag(Question question) {
+        List<TagDto> tagList = new ArrayList<>();
+        for (QuestionTag questionTag : question.getQuestionTags()) {
+            TagDto tagDto = new TagDto();
+            tagDto.setTagName(questionTag.getTag().getTagName());
+            tagDto.setTagDescription(questionTag.getTag().getTagDescription());
+            tagList.add(tagDto);
+        }
+
+        return tagList;
+    }
+
     public URI uriBuilder(long questionId, String basicURL) {
         return UriComponentsBuilder
                 .fromUriString(basicURL)
-                .path("/"+String.valueOf(questionId))
+                .path("/"+questionId)
                 .build()
                 .toUri();
     }
+    // --------------------------------------------------------------- 부기능
 
 
     // stub data zone
     public List<StubAnswer> stubAnswerList() {
         StubAnswer stubAnswer = new StubAnswer();
-        stubAnswer.setTitle("title");
-        stubAnswer.setDetail("detail");
+        stubAnswer.setAnswerTitle("title");
+        stubAnswer.setAnswerDetail("detail");
+        stubAnswer.setAnswerId(1L);
         return List.of(stubAnswer);
+    }
+
+    public MemberStub save(){
+        MemberStub memberStub = MemberStub.builder().name("member").build();
+        memberStubRepository.save(memberStub);
+        return memberStubRepository.findById(1L).orElseGet(null);
     }
